added HTTPClient Wifi only - move some logic to a new protocol
