+++ src/xcode/ENA/ENA/Source/Client/Client.swift
@@ -30,7 +30,6 @@ protocol Client {
 	typealias TestResultHandler = (Result<Int, Failure>) -> Void
 	typealias TANHandler = (Result<String, Failure>) -> Void
 	typealias DayCompletionHandler = (Result<SAPDownloadedPackage, Failure>) -> Void
-	typealias HourCompletionHandler = (Result<SAPDownloadedPackage, Failure>) -> Void
 	typealias CountryFetchCompletion = (Result<[Country], Failure>) -> Void
 
 	// MARK: Interacting with a Client
@@ -61,14 +60,6 @@ protocol Client {
 		_ day: String,
 		forCountry country: String,
 		completion: @escaping DayCompletionHandler
-	)
-
-	/// Fetches the keys for a given `hour` of a specific `day`.
-	func fetchHour(
-		_ hour: Int,
-		day: String,
-		country: String,
-		completion: @escaping HourCompletionHandler
 	)
 
 	// MARK: Getting the Configuration
@@ -170,7 +161,6 @@ struct FetchedDaysAndHours {
 }
 
 extension Client {
-	typealias FetchHoursCompletionHandler = (HoursResult) -> Void
 
 	/// Fetch the keys with the given days and country code
 	func fetchDays(
@@ -206,62 +196,4 @@ extension Client {
 		}
 	}
 
-	func fetchDays(
-		_ days: [String],
-		hours: [Int],
-		of day: String,
-		country: String,
-		completion completeWith: @escaping DaysAndHoursCompletionHandler
-	) {
-		let group = DispatchGroup()
-		var hoursResult = HoursResult(errors: [], bucketsByHour: [:], day: day)
-		var daysResult = DaysResult(errors: [], bucketsByDay: [:])
-
-		group.enter()
-		fetchDays(days, forCountry: country) { result in
-			daysResult = result
-			group.leave()
-		}
-
-		group.enter()
-		fetchHours(hours, day: day, country: country) { result in
-			hoursResult = result
-			group.leave()
-		}
-		group.notify(queue: .main) {
-			completeWith(FetchedDaysAndHours(hours: hoursResult, days: daysResult))
-		}
-	}
-
-	func fetchHours(
-		_ hours: [Int],
-		day: String,
-		country: String,
-		completion completeWith: @escaping FetchHoursCompletionHandler
-	) {
-		var errors = [Client.Failure]()
-		var buckets = [Int: SAPDownloadedPackage]()
-		let group = DispatchGroup()
-
-		hours.forEach { hour in
-			group.enter()
-			self.fetchHour(hour, day: day, country: country) { result in
-				switch result {
-				case let .success(hourBucket):
-					buckets[hour] = hourBucket
-				case let .failure(error):
-					errors.append(error)
-				}
-				group.leave()
-			}
-		}
-
-		group.notify(queue: .main) {
-			completeWith(
-				HoursResult(errors: errors, bucketsByHour: buckets, day: day)
-			)
-		}
-	}
-
-	typealias DaysAndHoursCompletionHandler = (FetchedDaysAndHours) -> Void
 }
